Add fileFilter

+++ Domain/ImageManager.cs
@@ -1,6 +1,7 @@
 using Model;
 using MetadataExtractor;
 using Directory = System.IO.Directory;
+using System.Text.RegularExpressions;
 
 namespace Domain;
 
@@ -8,15 +9,19 @@ public class ImageManager
 {
     private ImageFile? ImageFile { get; set; }
     const string pathImages = @"C:\Users\D13\Desktop\ImagesTest\";
-    const string imageName = "img";
+    const string imageName = "_img";
     const string folderOutput = "folder1";
     const string path = @$"{pathImages}";
     const string path2 = @$"{pathImages}{folderOutput}\";
+    
 
     public void GetAllImages()
     {
+
+        
+        string[] allowedExtensions = new [] {".jpeg", ".jpg", ".png"}; 
         //Read Path
-        foreach (string file in System.IO.Directory.EnumerateFiles(@$"{pathImages}", "*.jpg"))
+        foreach (string file in System.IO.Directory.EnumerateFiles(@$"{pathImages}").Where(file => allowedExtensions.Any(file.ToLower().EndsWith)).ToList())
         {
             //string contents = File.ReadAllText(file);
             var directories = ImageMetadataReader.ReadMetadata(file);
@@ -40,7 +45,12 @@ public class ImageManager
 
         try
         {
-            string[] filePaths = Directory.GetFiles(path, "*.jpg");
+            var searchPattern = new Regex(@"$(?<=\.(jpg|jpeg))");
+            
+            string[] allowedExtensions = new [] {".jpeg", ".jpg"}; 
+            //string[] filePaths = Directory.GetFiles(path, "*.jpg");
+            List<string> filePaths = Directory.GetFiles(@$"{pathImages}").Where(file => searchPattern.IsMatch(file)).ToList();
+            
             var counter = 1;
 
             DirectoryInfo directoryInfo = new DirectoryInfo(path);
@@ -49,13 +59,15 @@ public class ImageManager
 
             IEnumerable<FileInfo> filesSorted = files.OrderByDescending(x => x.LastWriteTime).ToList();
 
+            if (filesSorted.Count() == 0) return;
+
             foreach (var f in filesSorted)
             {
                 if (File.Exists(path2 + f.Name))
                     File.Delete(path2 + f.Name);
 
                 // Move the file.
-                File.Move(path + f.Name, path2 + $"{imageName}{counter}.jpg");
+                File.Move(path + f.Name, path2 + $"{counter}{imageName}{f.Extension}");
                 counter++;
             }
